rppr-f-1/requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart==0.0.6
bcrypt==4.0.1
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9